+++ system/modules/isotope/IsotopeFrontend.php
@@ -1317,6 +1317,22 @@ $endScript";
 		
 		return array_merge($arrPages, $arrIsotopeProductPages);
 	}
+	
+	
+	/**
+	 * save_callback for upload widget to store $_FILES data into the product
+	 * @param mixed
+	 * @param IsotopeProduct
+	 */
+	public function saveUpload($varValue, IsotopeProduct $objProduct, Widget $objWidget)
+	{
+		if (is_array($_SESSION['FILES'][$objWidget->name]) && $_SESSION['FILES'][$objWidget->name]['uploaded'] == '1' && $_SESSION['FILES'][$objWidget->name]['error'] == 0)
+		{
+			return $_SESSION['FILES'][$objWidget->name]['name'];
+		}
+
+		return $varValue;
+	}
 
 
 	/**