+++ system/modules/isotope/dca/tl_iso_products.php
@@ -2032,12 +2032,25 @@ $strBuffer .= '<th style="text-align:center"><img src="system/themes/default/ima
 			// Add date picker
 			if ($objAttributes->rgxp == 'date')
 			{
-				$arrData['eval']['datepicker'] = (method_exists($this,'getDatePickerString') ? $this->getDatePickerString() : true);
+				$arrData['eval']['datepicker'] = (method_exists($this, 'getDatePickerString') ? $this->getDatePickerString() : true);
 			}
 
+			// Textarea cannot be w50
 			if ($objAttributes->type == 'textarea' || $objAttributes->rte != '')
 			{
 				$arrData['eval']['tl_class'] = 'clr';
+			}
+			
+			// Customer defined widgets
+			if ($GLOBALS['ISO_ATTR'][$objAttributes->type]['customer_defined'])
+			{
+				$arrData['attributes']['customer_defined'] = true;
+			}
+			
+			// Install save_callback for upload widgets
+			if ($objAttributes->type == 'upload')
+			{
+				$arrData['save_callback'][] = array('IsotopeFrontend', 'saveUpload');
 			}
 
 			// Parse multiline/multilingual foreignKey